+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/utils/Constants.java
@@ -60,6 +60,16 @@ public class Constants {
     public static final String E_400_UNPROMOTABLE_SPACE = "Space [%s] cannot be promoted.";
     public static final String E_400_DUPLICATE_NAME =
             "A %s with the name '%s' already exists in the %s space.";
+    public static final String E_400_TOO_MANY_INTEGRATIONS =
+            "This request would create more than the allowed integrations [%d].";
+    public static final String E_400_TOO_MANY_DECODERS =
+            "This request would create more than the allowed decoders [%d].";
+    public static final String E_400_TOO_MANY_RULES =
+            "This request would create more than the allowed rules [%d].";
+    public static final String E_400_TOO_MANY_KVDBS =
+            "This request would create more than the allowed kvdbs [%d].";
+    public static final String E_400_TOO_MANY_FILTERS =
+            "This request would create more than the allowed filters [%d].";
     public static final String E_400_UUID_SHOULD_NOT_BE_PROVIDED =
             "ID should not be provided in the payload.";
     public static final String E_400_ENGINE_VALIDATION_FAILED = "Engine validation failed.";
@@ -88,6 +98,16 @@ public class Constants {
             "Unable to reach the CTI API to check for updates.";
 
     // Log messages
+    public static final String I_LOG_MAX_INTEGRATIONS_REACHED =
+            "This request would create more than the allowed integrations [{}].";
+    public static final String I_LOG_MAX_DECODERS_REACHED =
+            "This request would create more than the allowed decoders [{}].";
+    public static final String I_LOG_MAX_RULES_REACHED =
+            "This request would create more than the allowed rules [{}].";
+    public static final String I_LOG_MAX_KVDBS_REACHED =
+            "This request would create more than the allowed kvdbs [{}].";
+    public static final String I_LOG_MAX_FILTERS_REACHED =
+            "This request would create more than the allowed filters [{}].";
     public static final String I_LOG_SUCCESS = "{} {} successfully (id={})";
     public static final String D_LOG_OPERATION = "{} {} (id={})";
     public static final String E_LOG_ENGINE_IS_NULL = "Engine instance unavailable.";